setting: Replace parse implementations with custom derive macro [WIP]
